finetune/requirements.txt
absl-py==1.0.0
addict==2.4.0
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
anyio==3.6.2
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
astor==0.1
astunparse==1.6.3
attrs==19.1.0
baidu-aip==4.16.10
beautifulsoup4==4.11.1
certifi==2025.1.31
cffi==1.15.1
charset-normalizer==3.4.1
cityscapesScripts==2.2.4
click==8.1.8
codecov==2.1.13
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.1.1
contrib==0.3.0
coverage==7.6.1
crcmod==1.7
cryptography==44.0.2
cycler==0.12.1
decorator==5.2.1
exceptiongroup==1.2.2
ffmpeg-python==0.2.0
filelock==3.14.0
flake8==7.1.2
Flask==2.2.2
Flask-SocketIO==5.2.0
Flask-SQLAlchemy==3.0.3
flatbuffers==2.0
fonttools==4.56.0
ftfy==6.2.3
gast==0.6.0
google==3.0.0
google-auth==2.16.0
google-auth-oauthlib==0.4.6
grad-cam==1.3.7
h5py==3.8.0
handright==8.1.0
huggingface-hub==0.12.1
humanfriendly==10.0
idna==3.10
imagecodecs==2023.3.16
imageio==2.22.4
imageio-ffmpeg==0.2.0
importlib-resources==5.10.1
importlib_metadata==8.5.0
iniconfig==2.0.0
interrogate==1.7.0
ipykernel==6.17.1
ipython==8.7.0
ipywidgets==8.0.3
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.5
jmespath==0.10.0
joblib==1.0.0
jsonschema==4.0.0
jupyter-console==6.4.4
jupyter-events==0.5.0
jupyter_client==7.4.8
jupyter_core==5.1.0
jupyter_server==2.0.0
jupyter_server_terminals==0.4.2
kiwisolver==1.4.7
Markdown==3.4.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.6
mccabe==0.7.0
mdurl==0.1.2
mediapipe==0.9.2.1
mmcv==2.2.0
mmengine==0.10.7
mmsegmentation==1.2.2
model-index==0.1.11
moviepy==1.0.3
mpmath==1.3.0
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.2.6
nbformat==5.7.0
networkx==2.2
nibabel==5.2.1
notebook_shim==0.2.2
numpy==1.24.4
opencv-contrib-python==4.6.0.66
opencv-python==4.11.0.86
opendatalab==0.0.10
openmim==0.3.9
openxlab==0.1.2
opt-einsum==3.3.0
ordered-set==4.1.0
oss2==2.17.0
packaging==24.2
paddle-bfloat==0.1.7
paddlepaddle==2.4.2
pandas==2.0.3
pillow==10.4.0
platformdirs==4.3.6
pluggy==1.5.0
pretrainedmodels==0.7.4
prettytable==3.11.0
protobuf==3.20.0
py==1.11.0
pycodestyle==2.12.1
pycparser==2.22
pycryptodome==3.21.0
pyflakes==3.2.0
Pygments==2.19.1
pyinstaller==5.7.0
pyparsing==3.1.4
pyqt5-plugins==5.15.9.2.3
pyqt5-tools==5.15.9.3.3
pyquaternion==0.9.9
pyreadline3==3.5.4
pyrsistent==0.20.0
pytest==8.3.5
python-dateutil==2.9.0.post0
pytz==2023.4
PyWavelets==1.1.1
pywin32==308
PyYAML==6.0.2
qtconsole==5.4.2
regex==2024.11.6
requests==2.28.2
requests-oauthlib==1.3.1
rich==13.4.2
scikit-image==0.19.3
scikit-learn==1.1.3
scipy==1.10.1
seaborn==0.12.2
selenium==4.7.2
six==1.17.0
sounddevice==0.4.6
sympy==1.13.3
tabulate==0.9.0
tensorboard==2.11.2
tensorflow==2.11.0
tensorflow-intel==2.11.0
termcolor==2.4.0
tflite-support==0.4.3
thop==0.1.1.post2209072238
threadpoolctl==2.0.0
tifffile==2019.7.26
tomli==2.2.1
torchaudio==2.0.2
torchvision==0.15.2
tqdm==4.65.2
transformers==4.28.1
trio==0.22.0
trio-websocket==0.9.2
ttach==0.0.3
typing==3.7.4.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==1.26.20
wcwidth==0.2.13
wrapt==1.17.2
xdoctest==1.2.0
yapf==0.43.0
zipp==3.20.2
